requirements.txt
numpy>=1.17.2
pandas>=1.0.4
pillow>=7.2.0
matplotlib>=3.4.2